Added relationship from trackeditem to user. Updated all mongoose interacitons with trackedItems to check user id for authoriztion

+++ routes/TrackedItemRoutes.js
@@ -1,10 +1,13 @@
 const routes = require('express').Router();
 const trackedItemController = require('../controllers/TrackedItemController');
+const { checkLoggedIn }  = require('../middleware/auth');
 
-
-//Post a new item to track (link, phone number, target price)
-routes.post( '/trackedItem', async (req, res )=> {
+routes.post( '/trackedItem', checkLoggedIn, async (req, res )=> {
     const trackedItem = req.body;
+
+    const userId = req.user._id;
+    trackedItem.user = userId;
+
     const addedTrackedItem = await trackedItemController.addTrackedItem(trackedItem);
     
     if(addedTrackedItem === null) {
@@ -15,8 +18,10 @@ routes.post( '/trackedItem', async (req, res )=> {
         
 });
 
-routes.get('/trackedItem', async (req, res) => {
-    const trackedItems = await trackedItemController.getTrackedItems();
+//Should only get ones that the user created
+routes.get('/trackedItem', checkLoggedIn, async (req, res) => {
+    const userId = req.user._id;
+    const trackedItems = await trackedItemController.getTrackedItemsByUser(userId);
 
     if(trackedItems === null) {
         return res.status(500).send({msg: "Failed to save tracked item" });
@@ -25,10 +30,16 @@ routes.get('/trackedItem', async (req, res) => {
     return res.send(trackedItems);
 });
 
-//get specific tracked item
-routes.get('/trackedItem/:id', async (req, res) => {
+//TODO: can only view ones that the user created
+routes.get('/trackedItem/:id', checkLoggedIn, async (req, res) => {
     const id = req.params.id;
     const trackedItem = await trackedItemController.getTrackedItem(id);
+
+    const userId = req.user._id;
+
+    if(trackedItem.user !== userId) {
+        return res.status(403).send({msg: "Unauthorized"});
+    }
 
     if(trackedItem === null) {
         return res.status(404).send({msg: "Failed to find tracked item" });
@@ -37,20 +48,38 @@ routes.get('/trackedItem/:id', async (req, res) => {
     return res.send(trackedItem);
 });
 
-//update specfiic tracked item
-routes.put('/trackedItem/:id', (req, res) => {
+//TODO: can only update ones that the user created
+routes.put('/trackedItem/:id', checkLoggedIn, (req, res) => {
     const trackedItem = req.body;
     const id = req.params.id;
+
+    const trackedItemBefore = trackedItemController.getTrackedItem(id);
+    const userId = req.user._id;
+
+
+    if(trackedItem.user !== userId) {
+        return res.status(403).send({msg: "Unauthorized"});
+    }
     
     const updatedTrackedItem = trackedItemController.updateTrackedItem(id, trackedItem);
+
     if(updatedTrackedItem === null ){
         return res.status(404).send({msg: "Failed to find and update tracked item" });
     }
     return res.send(updatedTrackedItem);
 })
 
-routes.delete('/trackedItem/:id', async (req, res) => {
+//TODO: can only delete ones that the user created
+routes.delete('/trackedItem/:id', checkLoggedIn, async (req, res) => {
     const id = req.params.id;
+
+    const trackedItemBefore = trackedItemController.getTrackedItem(id);
+    const userId = req.user._id;
+
+    if(trackedItem.user !== userId) {
+        return res.status(403).send({msg: "Unauthorized"});
+    }
+
     const deletedItem = await trackedItemController.deleteTrackedItem(id);
 
     if(deletedItem === null ){
